server/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pydantic==2.6.4
mysql-connector-python==8.0.33
langgraph==0.0.39
langchain==0.1.16
langchain-community==0.0.36
langchain-groq
httpx==0.27.0
python-dotenv==1.0.1
sqlalchemy==2.0.25